requirements.txt
openai
sentence-transformers>=2.2.2
datasets
accelerate
llama-index>=0.9.0
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-finetuning
llama-index-embeddings-langchain
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-embeddings-huggingface-api
llama-index-llms-groq
torch>=2.0.0
tqdm
pandas>=1.3.0
PyPDF2>=3.0.0
nest-asyncio
python-dotenv
gradio>=4.0.0
scikit-learn
pathlib
jupyter